requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-filter>=23.2
requests>=2.31.0
python-decouple>=3.8
django-cors-headers>=4.3.0
Pillow>=10.0.0
boto3>=1.34.0
django-q2>=1.6.0
gunicorn
whitenoise
dj-database-url
psycopg2-binary
